feat: added validation to delete a clothe item

+++ src/main/java/com/leadtech/lookbook/service/ClothingItemService.java
@@ -37,7 +37,12 @@ public class ClothingItemService {
     }
 
     public void deletar(Long id) {
+        ClothingItem clothingItem = buscarPorId(id);
+        if (!clothingItem.getLookbooks().isEmpty()) {
+            throw new RuntimeException("A peça está vinculada a um lookbook");
+        }
         clothingItemRepository.deleteById(id);
     }
+
 }
